Refactor Unit Service Tests
Remove @BeforeEach setup methods for individual service tests and move
it to the parent class for better readabilitRefactor Unit Service Tests

Remove @BeforeEach setup methods for individual service
