+++ src/test/java/com/example/demo/unit/service/PoliticianServiceTest.java
@@ -1,6 +1,5 @@
 package com.example.demo.unit.service;
 
-import static com.example.demo.baseClasses.AbstractPoliticianControllerTest.withRepoAndId;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -9,11 +8,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
@@ -22,35 +18,9 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import com.example.demo.baseClasses.AbstractEntitiesServiceTest;
 import com.example.demo.dtoRequest.AddPoliticianDTORequest;
 import com.example.demo.model.entities.Politicians;
-import com.example.demo.model.entities.PoliticiansRating;
-import com.example.demo.model.entities.Rating;
-import com.example.demo.model.entities.UserRater;
-import com.example.demo.model.enums.PoliticalParty;
-import com.example.demo.service.PoliticiansService;
 
 @ExtendWith(SpringExtension.class)
 public class PoliticianServiceTest extends AbstractEntitiesServiceTest{
-	
-	@BeforeEach
-	public void setup() {
-		politicianService = new PoliticiansService(politicianRepo);
-		
-		List<PoliticiansRating> listOfPoliticiansRating = new ArrayList<>();
-		politician =  withRepoAndId
-				(ratingRepo,
-				1,
-				"Mirriam", 
-				"Defensor", 
-				listOfPoliticiansRating,
-				new Rating(0.01D, 0.01D, calculator),
-				"1");
-		
-		rating = new PoliticiansRating
-				(1, 8.023D, 
-				new UserRater("test", PoliticalParty.DDS, "test@gmail.com"),
-				politician);
-		politician.calculateListOfRaters(rating);
-	}
 	
 	@Test
 	public void verifyRepoCalledSaveMethod() {
@@ -90,7 +60,7 @@ public class PoliticianServiceTest extends AbstractEntitiesServiceTest{
 		when(politicianRepo.save(any(Politicians.class))).thenReturn(politician); 
 		when(politicianRepo.countByLastNameAndFirstName("Mirriam", "Defensor")).thenReturn(1L);
 		
-		Politicians politicianQueried = politicianService.savePolitician(new AddPoliticianDTORequest("Mirriam", "Defensor", BigDecimal.valueOf(0.01D)));
+		Politicians politicianQueried = politicianService.savePolitician(politicianDtoRequest);
 		
 		assertDtoOutputsUtil(politician, politicianQueried);
 	}
